backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
websockets==14.1
pydantic==2.10.3
python-multipart==0.0.12
aiosqlite==0.20.0
asyncpg==0.30.0
libsql-client>=0.0.1
# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0